CubesControl: Starting implementation of protocol level
ICubesProtocol _may_ be used in the future as an interface for various
protocols we may use. For now it seems like too much of a hassle with
my current C++ experience level to implement ICubesProtocol as an interface
to be used by implemented protocols.

For now, we will therefore just use CubesProtoUartPmod and add to it as
necessary.

When the time comes, ICubesProtocol should be used to:
- define interface methods for classes implementing the protocol
- define signals and slots common to the protocols

NB: no definition should be provided for data members of inside
ICubesProtocol. Members should be defined as part of the implementing
classes.