Fixed and tested writing register value on bit field change
Previously, the data byte order on serial port transfers was wrong; SIPHRA_CSR
was in fact written to as SIPHRA_DATAR and viceversa, hence the SPI transfer was
wrong. This was changed by changing the data byte order and it has been tested
to work (on the oscilloscope, at least).