CubesControl: Read All SIPHRA registers inside on_cubes_devReadReady()
The thread implementation in the previous commit did not work, perhaps due to
thread affinity of the CUBES objects. This is more a hunch and has not been
properly tested for confirmation. If the thread implementation is to be followed
later on, one should find a way to connect the CubesControl::on_cubes_devReadReady()
slot to the thread's cubes->devReadReady() signal. Again, this is a hunch, but
also the proposed way of testing whether the first hunch was right.

For now, doing all register reads and writes from on_cubes_devReadReady() works
fine, as the serial port communication is slow enough to allow the FPGA to
finish an SPI access to the SIPHRA.