CubesControl: Removed writing SIPHRA register on change to register bit field
This was deemed to be unsafe for now, until the ASIC has been fully tested, as in
some cases, for example setting the shaper via the CHANNEL_CONTROL register, there
are only specific settings allowed according to the ASIC datasheet.

Writing a setting that is unsupported by the ASIC when making a change to only one
of the bit fields of (in this example) the CHANNEL_CONTROL register would lead to an
undefined behavior according to the datasheet. It is safer to not do this until setting
these registers has been tested.