Updating histogram and ADC value label on new ADC value from FPGA
The updates to the UI widgets are made only if the user is viewing these
widgets, that is, if the tabWidget's current index is that which contains
the UI widget to be updated.

The histogram is updated with new data when the user switches to its tab.
This is a bit slow at the moment, to be fixed in some nice way...