CubesControl: Properly updating lblHistogramStatus
There was a small bug in the updating of this label when in time-based
histogramming mode. The on_btnStartStopHistogram_clicked() function was called
before doing the updating of the histogram label from the tmrHistogramAdcPoll
timeout callback; because on_btnStartStopHistogram_clicked() (the fcn. that
updates the label to its final state) was called before the update of the label
from the timeout callback, there was one more update of the label to
"Histogram running for 0 s."

By placing the call of the button clicked callback at the end of the timeout
callback, the issue has been fixed.