Warning the user when enabling ADC_IN use from diagnostics tab
Because of current limitations in the gateware and software, when using
the ADC_IN pin for checking the ADC range, the user should only set the
adc_in bit in READOUT_FIXED_LIST. If otherwise, it may be that the label
containing the current ADC value readout will contain a wrong value due
to either the FPGA storing another channel than the ADC_IN channel in its
ADCR, or due to the UART buffer containing the readout for another channel.

The user is thus warned to clear all other bits set in READOUT_FIXED_LIST
and keep only the ADC_IN bit. This allows for correct updating of the ADC
value label with the voltage at ADC_IN.