Using SIPHRA register write action from visual config
The on_actionWriteSiphraRegister_triggered() callback is used to perform
an actual write to the SIPHRA via the FPGA when a register is written
using the SIPHRA visual config tab.

This is more of a tweak to the program, as the functionality implemented
in the functions called when controlling UI elements was already implemented
as an action.