Correctly send I2C address in CubesProtoUartPmod.writeI2C
Due to the bitwise OR-ing of the CUBES I2C address with the current
state of the boolean write, the I2C address + R/nW byte was wrong.

In I2C, the ADDR+R/W byte has the following format:

bits 7..1  ADDR
bit     0  R/nW ('1' = read / '0' = write)

In the current version of the code, the boolean variable 'write' is
used in lieu of this bit. When 'write' is true, an I2C write should
occur, thus the write bit should be '0'. Inside the writeI2C function,
the boolean variable was negated bitwise, thus its value after bitwise
negation was 0xfe. Then, this value was ORed together with the I2C
address of CUBES, currently 0x70, which yielded again 0xfe. Thus, the
slave would not reply.

The implemeted fix corrects this issue and the slave can be addressed.