Remove unnecessary slot for QSerialPort::readyRead()
In the CubesProtoUartPmod class, the on_serialPort_readyRead() slot was
used in the previous implementation of the class for handling reads from
the serial port. In the current implementation, we may not use it any
more and it has been removed as a slot for the SerialPort::readyRead
signal, so it doesn't inadvertently respond to the signal and cause
unexpected behavior.