Change gen_asic_probes output file name
The file name expected by the Arduino file system handling function is
"8.3", i.e., 8 chars for the name and 3 for the extension. The gen-asic-probes
program was incompatible with this format, as it was generating a file called
"Probe.conf".

The output file name is now "probe.cfg", which matches with the 8.3 format.