Fix file naming and access type in file converter script
1. The config file has been renamed to "config.cfg" from "config.txt",
to reflect it is a configuration file rather than a text file.
2. The access type for the configuration file has been changed to "w+b",
which opens the file for writing and reading in binary mode, as well as
creating the file if it doesn't exist. The previous access type ("r+b")
was not creating the file if it did not exist, which lead to the script
failing.